This diff aims to reduce the size of the `Value` struct from 80 bytes to 32 bytes by storing a unique pointer to `StagingBuffer` and `SymInt` instead of inlining the objects.

Differential Revision: [D66677759](https://our.internmc.facebook.com/intern/diff/D66677759/)

[ghstack-poisoned]

+++ backends/vulkan/runtime/api/containers/Tensor.h
@@ -317,7 +317,7 @@ class vTensor final {
   // each uniform is assumed to be a vec of 4 ints to maintain 16 byte alignemnt
   constexpr static size_t kSizePerUniform = sizeof(utils::ivec4);
   // Total size of tensor's uniform buffer
-  constexpr static uint32_t kMaxUniformBufferSize =
+  constexpr static size_t kMaxUniformBufferSize =
       4 * // we have 4 uniforms that are passed on to shaders
       kSizePerUniform;
